remove Win32InputBox.cpp/.h, move them to UIEditBoxImpl-win32.cpp
